requirements.txt
# Библиотеки для работы с криптобиржами
python-binance>=1.0.19
pybit>=5.6.0

# Для работы с переменными окружения
python-dotenv>=1.0.0
